+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Other: Introduce SQLAlchemy dialect compliance test suite and enumerate all excluded tests
 - Add integration tests for Databricks UC Volumes ingestion queries
+- Add `_retry_max_redirects` config
 
 ## 2.9.3 (2023-08-24)